refactor: layout과 metadata 분리 및 Toaster 위치/속성 개선 (#114)
* feat: metadata, viewport 설정 파일 분리

* refactor: metadata, viewport 분리 및 Toaster 위치 조정
